Send emails on file download, upload, delete with SES

+++ src/main/java/com/valentin/demo_aws/controller/FileController.java
@@ -1,6 +1,7 @@
 package com.valentin.demo_aws.controller;
 
 import com.valentin.demo_aws.model.FileMetadata;
+import com.valentin.demo_aws.service.EmailService;
 import com.valentin.demo_aws.service.FileService;
 import com.valentin.demo_aws.service.S3Service;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +22,8 @@ public class FileController {
     private FileService fileService;
     @Autowired
     private S3Service s3Service;
+    @Autowired
+    private EmailService emailService;
 
     @GetMapping
     public ResponseEntity<List<String>> listFiles() {
@@ -35,6 +38,10 @@ public class FileController {
             return ResponseEntity.notFound().build();
         }
 
+        FileMetadata fileMetadata = fileOpt.get();
+        String currentUser = "request@author.com";
+        emailService.sendDownloadNotification(fileMetadata, currentUser);
+
         String url = s3Service.generateFileUrl(fileOpt.get().getKey());
         return ResponseEntity.status(HttpStatus.FOUND)
                 .location(URI.create(url))
@@ -44,10 +51,15 @@ public class FileController {
     @PostMapping("/upload")
     public ResponseEntity<String> uploadFile(
             @RequestParam("file") MultipartFile file,
-            @RequestParam("description") String description) {
+            @RequestParam("description") String description,
+            @RequestParam("uploader-email") String uploaderEmail) {
 
         try {
-            fileService.uploadFile(file, file.getOriginalFilename(), description);
+            FileMetadata fileMetadata = fileService.uploadFile(file, file.getOriginalFilename(), description, uploaderEmail);
+
+            String currentUser = "request@author.com";
+            emailService.sendUploadConfirmationEmail(fileMetadata, currentUser);
+
             return ResponseEntity.status(HttpStatus.OK)
                     .body("File uploaded successfully");
         } catch (Exception e) {
@@ -59,6 +71,15 @@ public class FileController {
     @DeleteMapping("/delete")
     public ResponseEntity<String> deleteFile(@RequestParam String key) {
         try {
+            Optional<FileMetadata> fileOpt = fileService.findFileByKey(key);
+            if (fileOpt.isEmpty()) {
+                return ResponseEntity.notFound().build();
+            }
+
+            FileMetadata fileMetadata = fileOpt.get();
+            String currentUser = "request@author.com";
+            emailService.sendDeleteNotification(fileMetadata, currentUser);
+
             fileService.deleteFile(key);
             return ResponseEntity.ok("File deleted successfully");
         } catch (Exception e) {
